+++ src/lib/templates/mod.rs
@@ -3,3 +3,4 @@
 pub mod defaults;
 pub mod manager;
 pub mod template;
+pub mod utils;